pyproject.toml
[project]
version = "0.2.0"
authors = [{ email = "sbarrios93@gmail.com" }, { name = "Sebastian Barrios" }]
description = "Amazon YNAB is a reconciler that scrapes Amazon orders and adds memo info on each corresponding YNAB transaction"
license = { text = "MIT" }
name = "amazon-ynab"
readme = "README.md"
requires-python = ">=3.10.0,<4.0.0"
dependencies = [
    "beautifulsoup4~=4.12.2",
    "nltk~=3.8.1",
    "selenium~=4.9.0",
    "webdriver-manager~=3.8.6",
    "typer[all]~=0.7.0",
    "rich~=12.6.0",
    "pyyaml~=6.0",
]

[tool.rye]
dev-dependencies = [
    "types-PyYAML~=6.0.12.9",
    "black~=23.3.0",
    "types-selenium~=3.141.9",
    "pyright~=1.1.304",
    "ruff~=0.0.262",
    "pytest~=7.3.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.black]
# https://github.com/psf/black
color = true
line-length = 88
target-version = ["py311"]
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''
preview = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
doctest_optionflags = [
    "NUMBER",
    "NORMALIZE_WHITESPACE",
    "IGNORE_EXCEPTION_DETAIL",
]
norecursedirs = [
    "hooks",
    "*.egg",
    ".eggs",
    "dist",
    "build",
    "docs",
    ".tox",
    ".git",
    "__pycache__",
]

# Extra options:
addopts = [
    "--strict-markers",
    "--tb=short",
    "--doctest-modules",
    "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[tool.pyright]
typeCheckingMode = "basic"
